New script to_pathologic.py for topf (issue #30). Fix issue in topf with pf files.
