+++ mpwt/__init__.py
@@ -1,3 +1,4 @@
 from mpwt.pwt_wrapper import run_pwt, run_pwt_dat
 from mpwt.mpwt_workflow import multiprocess_pwt
-from mpwt.utils import cleaning, cleaning_input, create_pathologic_file, find_ptools_path, list_pgdb, pubmed_citations, remove_pgdbs
+from mpwt.utils import cleaning, cleaning_input, find_ptools_path, list_pgdb, pubmed_citations, remove_pgdbs
+from mpwt.to_pathologic import create_pathologic_file
